Title: A, B, C 찾기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241029/A, B, C 찾기/finding-a-b-c.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 int values[7];
 int main() {
-    int a, b, c;
+    long int a, b, c;
 
     for(int i = 0; i < 7; i++){
         cin>> values[i];
